Got tired of the same block everywhere, now changes on height

+++ Assets/Scripts/Cell.cs
@@ -1,14 +1,17 @@
 
-public class Cell {
-    public bool isWater;
-    public int height;
+public class Cell
+{
+  public bool isWater;
+  public float height;
 
-    public Cell(bool isWater) {
-        this.isWater = isWater;
-    }
+  public Cell(bool isWater)
+  {
+    this.isWater = isWater;
+  }
 
-    public Cell(bool isWater, int height) {
-        this.isWater = isWater;
-        this.height = height;
-    }
+  public Cell(bool isWater, float height)
+  {
+    this.isWater = isWater;
+    this.height = height;
+  }
 }
